updateExposure: allow exceptions if data doesn't exist
The various functions to retrieve and apply the correction data
were silently proceeding in the event that the data did not exist,
misleading the user. Instead, allow exceptions to bubble up ---
the user asked us to retrieve and/or apply the correction, so if
we can't do that there's a problem that the user needs to know
about.

Ported from LSST; resolved conflict due to lack of support in HSC
for the old HSC pipeline.